Change UI when connected status change
When a person is connected we disable the
account creation and the "log in" page and
we show the "account" and "log-out" pages.
When the user is disconnected, the "sign-in"
and "Login" pages are here.
A login service provide a better way to handle
the fact that a user can be connected.